jwt auth done

+++ backend/controllers/goalController.js
@@ -1,23 +1,26 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../model/goalModel');
+const User = require('../model/userModel');
 
+const getGoals = asyncHandler(async (req, res) => {
+  const goals = await Goal.find({ user: req.user.id })
 
-const getGoals = asyncHandler(async(req, res) => {
-    const goals = await Goal.find();
-    res.status(200).json( goals);
+  res.status(200).json(goals)
 })
 
-const setGoal = asyncHandler(async(req, res) => {
-   if (!req.body.text){
+const setGoal = asyncHandler(async (req, res) => {
+  if (!req.body.text) {
     res.status(400)
-    throw new Error('Please Enter  Text field');
-   }
+    throw new Error('Please add a text field')
+  }
 
-   const goal = await Goal.create({
-         text:req.body.text,
-    })
-    res.status(200).json(goal);
+  const goal = await Goal.create({
+    text: req.body.text,
+    user: req.user.id,
+  })
+
+  res.status(200).json(goal)
 })
 
 
@@ -26,6 +29,20 @@ const updateGoal = asyncHandler(async(req, res) => {
     if(!goal){
         res.status(404)
         throw new Error('Goal not found');
+    }
+
+    const user = await User.findById(req.user._id);
+
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found');
+    }
+
+    //check for goal
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized');
     }
  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,req.body,{
         new:true,
@@ -39,8 +56,22 @@ const deleteGoal = asyncHandler(async(req, res) => {
         res.status(404)
         throw new Error('Goal not found');
     } 
+    const user = await User.findById(req.user._id);
+
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found');
+    }
+
+    //check for goal
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized');
+    }
 
     await goal.deleteOne();
+    
     res.status(200).json({id: req.params.id});
 })
 module.exports = { getGoals, setGoal, updateGoal, deleteGoal };
